Восстановление обработчиков попапа и позици прокрутки
