+++ src/presenter/Film.js
@@ -1,6 +1,5 @@
 import FilmCardView from '../view/film-card.js';
 import FilmCardPopupView from '../view/card-popup.js';
-import NewCommentFormView from '../view/new-comment-form.js';
 import {render, RenderPosition, remove, replace} from '../utils/render.js';
 import {onEscKeyDown} from '../utils/common.js';
 
@@ -46,16 +45,12 @@ export default class FilmPresenter {
 
     if (prevPopupComponent) {
       replace(this._popupComponent, prevPopupComponent);
+
+      this._setScrollPosition();
     }
 
     remove(prevFilmComponent);
     remove(prevPopupComponent);
-  }
-
-  _renderNewCommentFrom() {
-    const newCommentForm = new NewCommentFormView();
-    const newCommentFormContainer = this._popupComponent.getElement().querySelector(`.film-details__bottom-container`);
-    render(newCommentFormContainer, newCommentForm, RenderPosition.BEFOREEND);
   }
 
   _popupOpenHandler(evt) {
@@ -65,7 +60,7 @@ export default class FilmPresenter {
 
       render(document.body, this._popupComponent, RenderPosition.BEFOREEND);
 
-      this._renderNewCommentFrom();
+      this._popupComponent.restoreHandlers();
 
       document.body.classList.add(`hide-overflow`);
       document.body.addEventListener(`click`, this._popupCloseHandler);
@@ -76,6 +71,7 @@ export default class FilmPresenter {
   _popupCloseHandler(evt) {
     if (evt.target.matches(`.film-details__close-btn`) || onEscKeyDown(evt)) {
       this.resetView();
+      this._popupComponent.reset();
 
       document.body.classList.remove(`hide-overflow`);
       document.body.removeEventListener(`click`, this._popupCloseHandler);
@@ -83,7 +79,17 @@ export default class FilmPresenter {
     }
   }
 
+  _getScrollPosition() {
+    this._scrollPosition = this._popupComponent.getElement().scrollTop;
+  }
+
+  _setScrollPosition() {
+    this._popupComponent.getElement().scrollTo(0, this._scrollPosition);
+  }
+
   _controlsClickHandler(evt) {
+    this._getScrollPosition();
+
     if (evt.target.matches(`.film-card__controls-item--add-to-watchlist`) || evt.target.id === `watchlist`) {
       this._changeData(Object.assign({}, this._film, {
         isInWatchList: !this._film.isInWatchList
